Merge in COSMOSEE/base from scripts_dev to master

* commit '172e2a139106b1a82ac39ec043d9bb64f85dea8b':
  Make scripts take 'dev' option

+++ scripts/linux/cosmos_cleanup.sh
@@ -3,11 +3,13 @@
 read -p "This will remove all docker volumes which will delete **ALL** stored commands and telemetry! Are you sure (Y/[N])? " choice
 case "$choice" in
   y|Y )
-  docker volume rm cosmos-elasticsearch-v
-  docker volume rm cosmos-grafana-v
   docker volume rm cosmos-minio-v
   docker volume rm cosmos-redis-v
   docker volume rm cosmos-gems-v
+  if [[ "$1" == "dev" ]]; then
+    docker volume rm cosmos-elasticsearch-v
+    docker volume rm cosmos-grafana-v
+  fi
   docker network rm cosmos
   ;;
   * )
